Fix next.config.js with es2015 babel preset
