+++ package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "babel-preset-es2015": "^6.24.1",
     "next": "4.0.1",
     "nextein": "1.0.0-beta24",
     "react": "16.0.0",
